+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - API: Renamed `error` to `task_error`
 - API: Return validation error in 400 body
+- BUG: Store scheduled dependencies
 
 # 1.61.0 (2025-05-05)